refactor(cli): remove parser interface with single implementation (#6486)
A long long time ago we had both an esbuild and tree-sitter parser,
eventually moving to tree-sitter and dropping the esbuild one. With only
a single parser there is no need for the interface and extra logic to
separate tests from that one implementation etc.

---

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases

GitOrigin-RevId: ad5cfda8f0accc61bcb6072a7cd59e3f70f80b61
