+++ gazelle/js/BUILD.bazel
@@ -26,7 +26,6 @@ go_library(
         "//gazelle/common/starlark",
         "//gazelle/js/node",
         "//gazelle/js/parser",
-        "//gazelle/js/parser/treesitter",
         "//gazelle/js/pnpm",
         "//gazelle/js/proto",
         "//gazelle/js/typescript",
