+++ gazelle/js/parser/BUILD.bazel
@@ -1,8 +1,19 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "parser",
     srcs = ["parser.go"],
     importpath = "aspect.build/cli/gazelle/js/parser",
     visibility = ["//visibility:public"],
+    deps = [
+        "//gazelle/common/treesitter",
+        "//pkg/logger",
+        "@com_github_smacker_go_tree_sitter//:go-tree-sitter",
+    ],
+)
+
+go_test(
+    name = "parser_test",
+    srcs = ["parser_test.go"],
+    embed = [":parser"],
 )
