+++ gazelle/js/parser/parser_test.go
@@ -1,9 +1,7 @@
-package tests
+package parser
 
 import (
 	"testing"
-
-	"aspect.build/cli/gazelle/js/parser"
 )
 
 var testCases = []struct {
@@ -11,7 +9,6 @@ var testCases = []struct {
 	filename        string
 	expectedImports []string
 	expectedModules []string
-	typeOnly        bool
 }{
 	{
 		desc:     "empty",
@@ -219,7 +216,6 @@ var testCases = []struct {
 		`,
 		filename:        "sg-example-once-crashed.tsx",
 		expectedImports: []string{"react"},
-		typeOnly:        true,
 	},
 	{
 		desc: "ts type import",
@@ -229,7 +225,6 @@ var testCases = []struct {
 		`,
 		filename:        "types.ts",
 		expectedImports: []string{"react", "y"},
-		typeOnly:        true,
 	},
 	{
 		desc: "include imports only used as types",
@@ -239,7 +234,6 @@ var testCases = []struct {
 		`,
 		filename:        "typeImport.ts",
 		expectedImports: []string{"my/types"},
-		typeOnly:        true,
 	},
 	{
 		desc: "include require()s only used as types",
@@ -249,7 +243,6 @@ var testCases = []struct {
 		`,
 		filename:        "typeRequire.ts",
 		expectedImports: []string{"my/types"},
-		typeOnly:        true,
 	},
 	{
 		desc: "include type-only imports",
@@ -259,7 +252,6 @@ var testCases = []struct {
 		`,
 		filename:        "typeImport.ts",
 		expectedImports: []string{"my/types"},
-		typeOnly:        true,
 	},
 	{
 		desc: "include unused type-only imports",
@@ -268,7 +260,6 @@ var testCases = []struct {
 		`,
 		filename:        "typeImport-unused.ts",
 		expectedImports: []string{"my/types"},
-		typeOnly:        true,
 	},
 	{
 		desc: "module declaration",
@@ -287,7 +278,6 @@ var testCases = []struct {
 		`,
 		filename:        "declare-module.ts",
 		expectedModules: []string{"module-x", "module-with-no-body", "comments-2", "comments-3", "module-quotes-1"},
-		typeOnly:        true,
 	},
 	{
 		desc: "declare module sub-imports",
@@ -304,7 +294,6 @@ var testCases = []struct {
 		filename:        "declare-module-sub.ts",
 		expectedModules: []string{"lib-imports"},
 		expectedImports: []string{"lib-export-star", "lib-export-star-as", "lib-from-default", "lib-impt"},
-		typeOnly:        true,
 	},
 	{
 		desc: "declare module protocol",
@@ -318,28 +307,10 @@ var testCases = []struct {
 		filename:        "declare-protocol-module.ts",
 		expectedModules: []string{"https://mod.com"},
 		expectedImports: []string{"ftp://ancient.com"},
-		typeOnly:        true,
 	},
 }
 
-func RunParserTests(t *testing.T, parser parser.Parser, includeTypeOnly bool, parserPost string) {
-	for _, tc := range testCases {
-		if !includeTypeOnly && tc.typeOnly {
-			continue
-		}
-
-		t.Run(tc.desc+parserPost, func(t *testing.T) {
-			res, _ := parser.ParseSource(tc.filename, tc.ts)
-
-			if !equal(res.Imports, tc.expectedImports) {
-				t.Errorf("Unexpected import results\nactual:  %#v;\nexpected: %#v\ntypescript code:\n%v", res.Imports, tc.expectedImports, tc.ts)
-			}
-
-			if !equal(res.Modules, tc.expectedModules) {
-				t.Errorf("Unexpected module results\nactual:  %#v;\nexpected: %#v\ntypescript code:\n%v", res.Modules, tc.expectedModules, tc.ts)
-			}
-		})
-	}
+func RunParserTests(t *testing.T, parserPost string) {
 }
 
 func equal[T comparable](a, b []T) bool {
@@ -352,4 +323,20 @@ func equal[T comparable](a, b []T) bool {
 		}
 	}
 	return true
+}
+
+func TestTreesitterParser(t *testing.T) {
+	for _, tc := range testCases {
+		t.Run(tc.desc, func(t *testing.T) {
+			res, _ := ParseSource(tc.filename, tc.ts)
+
+			if !equal(res.Imports, tc.expectedImports) {
+				t.Errorf("Unexpected import results\nactual:  %#v;\nexpected: %#v\ntypescript code:\n%v", res.Imports, tc.expectedImports, tc.ts)
+			}
+
+			if !equal(res.Modules, tc.expectedModules) {
+				t.Errorf("Unexpected module results\nactual:  %#v;\nexpected: %#v\ntypescript code:\n%v", res.Modules, tc.expectedModules, tc.ts)
+			}
+		})
+	}
 }
