--- gazelle/js/parser/tests/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "tests",
-    srcs = ["parser.go"],
-    importpath = "aspect.build/cli/gazelle/js/parser/tests",
-    visibility = ["//visibility:public"],
-    deps = ["//gazelle/js/parser"],
-)
