--- gazelle/js/parser/treesitter/parser_treesitter_test.go
@@ -1,11 +1,0 @@
-package gazelle
-
-import (
-	"testing"
-
-	"aspect.build/cli/gazelle/js/parser/tests"
-)
-
-func TestTreesitterParser(t *testing.T) {
-	tests.RunParserTests(t, NewParser(), true, "treesitter")
-}
